remove the security subsystem
Scuttlebutt's security subsystem is in fact fairly useless, because
updates can simply be left out by malicious nodes, and because updates
must arrive in-order, it is possible to cause past messages to be
ignored if they have not reached the local node yet.
The security subsystem unnecessarily complicates the code, and provides
a false sense of security. Scuttlebutt must be used in trusted
environments for it to function correctly. Thus, I'm removing the
security subsystem.